+++ CHANGELOG.md
@@ -81,6 +81,7 @@ Changes
 * [UI/Developer]: Updated analysis urls to work with a context path.
 * [UI]: Outputs and Tree view now span the full width of the page.
 * [Processing]: Fixed bug where analyses wouldn't run if project didn't have analysis priority.
+* [UI]: Updated the analysis results page to auto update the progress of an analysis.
 
 19.09 to 20.01
 --------------